+++ changelog.rst
@@ -1,14 +1,14 @@
 Changelog
 =========
 
-Last change: 07-DEC-2025 CEST
+Last change: 10-DEC-2025 CEST
 
 0.9.2
 -----
 Important updates:
 ^^^^^^^^^^^^^^^^^^
 - Improved and updated `sample notebooks <https://github.com/jungmannlab/picasso/tree/master/samples>`__.
-- Render: FRC resolution implementation
+- Render: FRC resolution implementation, see DOI: `10.1038/nmeth.2448 <https://doi.org/10.1038/nmeth.2448>`__
 
 *Small improvements:*
 +++++++++++++++++++++